% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{add.scu.flag}
\alias{add.scu.flag}
\title{Add Medeiros and Stevens (2011)-like Sc flag to file}
\usage{
add.scu.flag(fname = "../vignettes/EAMv1_entrain_NEP.cam.h1.2000_scu.nc")
}
\description{
Add Medeiros and Stevens (2011)-like Sc flag to file
}
