% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adiabatic_nd_from_tau_re}
\alias{adiabatic_nd_from_tau_re}
\title{Gamma is the adiabatic condensation rate in kg m^-4}
\usage{
adiabatic_nd_from_tau_re(tau, re, Gamma = 2e-06)
}
\description{
Gamma is the adiabatic condensation rate in kg m^-4
}
