% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R, R/process_e3sm.R, R/process_era5.R
\name{advection}
\alias{advection}
\title{Calculate horizontal scalar advection from u and v}
\usage{
advection(s, u, v, dx, dy)

advection(s, u, v, dx, dy)

advection(s, u, v, dx, dy)
}
\arguments{
\item{s}{Scalar quantity (2- or higher-dimensional array)}

\item{u}{Zonal wind component (2- or higher-dimensional array,
dimensionality matching scalar)}

\item{v}{Meridional wind component}

\item{dx}{Zonal grid spacing in units of m; two-dimensional (lon-lat)}

\item{dy}{Meridional grid spacing in units of m; one-dimensional (lat)}
}
\description{
Calculate horizontal scalar advection from u and v

Calculate horizontal scalar advection from u and v

Calculate horizontal scalar advection from u and v
}
