% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{calculate_budgets}
\alias{calculate_budgets}
\title{Calculate budgets}
\usage{
calculate_budgets(df)
}
\arguments{
\item{df}{data_frame grouped by all coordinate variables except lev}
}
\description{
Calculate budgets
}
