% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{calculate_h}
\alias{calculate_h}
\title{Calculate PBL top height}
\usage{
calculate_h(df, debug.theta = FALSE, debug.q = FALSE)
}
\arguments{
\item{df}{data_frame grouped by all coordinate variables except lev}
}
\description{
Calculate PBL top height
}
