% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{calculate_tendencies}
\alias{calculate_tendencies}
\title{Calculate tendencies}
\usage{
calculate_tendencies(df, lead = TRUE)
}
\arguments{
\item{df}{data_frame grouped by all coordinate variables except time}
}
\description{
Calculate tendencies
}
