% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{calculate_thermodynamics}
\alias{calculate_thermodynamics}
\title{Calculate thermodynamics (theta, theta_l, qt)}
\usage{
calculate_thermodynamics(df)
}
\arguments{
\item{df}{data.frame containing p, p0, t (temperature), qv, ql}
}
\description{
Calculate thermodynamics (theta, theta_l, qt)
}
