% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echam_utils.R
\name{canonize_fields_echam}
\alias{canonize_fields_echam}
\title{Replace certain fields with the canonical combinations used in our
chosen form of the budget equations}
\usage{
canonize_fields_echam(df)
}
\description{
Replace certain fields with the canonical combinations used in our
chosen form of the budget equations
}
