% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{cobble.on.precc}
\alias{cobble.on.precc}
\title{Cobble on PRECC, which is only included in the process.F2010A.entrain() output as of mltools >= 0.0.14}
\usage{
cobble.on.precc(
  path = "~/scratch/e3sm_scratch/v2_F2010A0_00/run",
  pattern = ".*.eam.h9..*.nc",
  time.min = "1990-01-01",
  time.max = "2010-01-01",
  h.entrain = "eam.h9",
  parallel = TRUE
)
}
\description{
Cobble on PRECC, which is only included in the process.F2010A.entrain() output as of mltools >= 0.0.14
}
