% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{convert_conventional_units}
\alias{convert_conventional_units}
\title{Convert to conventional LWP and Nd units}
\usage{
convert_conventional_units(df)
}
\description{
Convert to conventional LWP and Nd units
}
