% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\name{dist_fine_to_coarse}
\alias{dist_fine_to_coarse}
\title{Map ne120pg2 grid points to closest ne30pg points}
\usage{
dist_fine_to_coarse(fine, coarse, ...)
}
\description{
Map ne120pg2 grid points to closest ne30pg points
}
