% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_eamxx.R
\name{eamxx_reduce_parallel}
\alias{eamxx_reduce_parallel}
\title{Reduce fields stepwise}
\usage{
eamxx_reduce_parallel(nc, varnames.3d, varnames.2d = NULL, eamxx.steps, reduce)
}
\description{
Supply a file name instead of an ncdf4 object, since each thread will need
its own file handle
}
