% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_eamxx.R
\name{eamxx_step}
\alias{eamxx_step}
\title{Calculate start and count values for blockwise traversal of EAMxx (i.e., very
large) NetCDF files}
\usage{
eamxx_step(nc, varname, blk.time = 1, blk.ncol = 1024^2, blk.lev = 128)
}
\value{
data.frame with start and count for time, ncol, lev
}
\description{
Only intended for 3D fields with time, ncol, lev dimension for now
}
