% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_eamxx.R
\name{eamxx_step_ncol_select}
\alias{eamxx_step_ncol_select}
\title{Calculate start and count values for blockwise traversal of EAMxx (i.e., very
large) NetCDF files}
\usage{
eamxx_step_ncol_select(nc, varname, blk.time = 1, ncols, blk.lev = 128)
}
\arguments{
\item{ncols}{Vector of ncols to select; each entry turns into a size-1 block}
}
\value{
data.frame with start and count for time, ncol, lev
}
\description{
Only intended for 3D fields with time, ncol, lev dimension for now
}
