% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwp_utils.R
\name{filter_standard}
\alias{filter_standard}
\title{Apply all filters used in the standard analysis (warm cloud, ocean, Sc dynamical regime)}
\usage{
filter_standard(df)
}
\description{
Apply all filters used in the standard analysis (warm cloud, ocean, Sc dynamical regime)
}
