% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{find_inversion_eamxx}
\alias{find_inversion_eamxx}
\title{Find inversion}
\usage{
find_inversion_eamxx(df, default.type = "jk.tinv")
}
\arguments{
\item{df}{data_frame grouped by all coordinate variables except lev}

\item{default.type}{which inversion criterion gets returned as
jk.pbl; "factor" returns all, with inv.type indicating the inversion type}
}
\description{
Find inversion
}
