% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwp_utils.R
\name{inverted.v}
\alias{inverted.v}
\title{Parameterization of the "inverted v" Nd-LWP relationship}
\usage{
inverted.v(N, Np, Lp, ml, mh)
}
\arguments{
\item{N}{really log(N)}

\item{Np}{really the log of the location of the peak of the inverted v}

\item{Lp}{really the log of the location of the peak of the inverted v}

\item{ml}{the slope below the peak}

\item{mh}{the slope above the peak}
}
\value{
log(L)
}
\description{
Note: the parameterization is piecewise linear in log-log space; in
the notation below, the log is implicit
}
