% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwp_utils.R
\name{inverted_v_path_data}
\alias{inverted_v_path_data}
\title{Parameterization of the "inverted v" Nd-LWP relationship}
\usage{
inverted_v_path_data(Np, Lp, ml, mh, Nmin = 1e+06, Nmax = 3e+08, n.points = 2)
}
\arguments{
\item{Np}{really the log of the location of the peak of the inverted v}

\item{Lp}{really the log of the location of the peak of the inverted v}

\item{ml}{the slope below the peak}

\item{mh}{the slope above the peak}

\item{N}{really log(N)}
}
\value{
log(L)
}
\description{
Note: the parameterization is piecewise linear in log-log space; in
the notation below, the log is implicit
}
