% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwp_utils.R
\name{load_lwp_cam}
\alias{load_lwp_cam}
\title{Load the fields required for the LWP-CDNC analysis (Andrew Gettelman's CAM run)}
\usage{
load_lwp_cam(fname)
}
\description{
Load the fields required for the LWP-CDNC analysis (Andrew Gettelman's CAM run)
}
