% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\name{map_fine_to_coarse}
\alias{map_fine_to_coarse}
\title{Map ne120pg2 grid points to closest ne30pg2 points}
\usage{
map_fine_to_coarse(fine, coarse, dist = plotutils::dist.gc, ...)
}
\arguments{
\item{dist}{distance metric between grid points; defaults to
plotutils::dist.gc(), expected signature is (lon1, lon2, lat1, lat2)}

\item{...}{arguments passed to ddply}
}
\description{
Map ne120pg2 grid points to closest ne30pg2 points
}
\examples{
df.120.map <- map_fine_to_coarse(df.120, df.30, .parallel = TRUE, .progress = "text")
df.120.map <- map_fine_to_coarse(df.120, df.30,
    dist = function(lon1, lon2, lat1, lat2) sqrt((lon1 - lon2)^2 + (lat1 - lat2)^2), 
    .parallel = TRUE, .progress = "text")
}
