% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\docType{package}
\name{mltools-package}
\alias{mltools}
\alias{mltools-package}
\title{Tools for analyzing (Scu-like) boundary layers in GCMs}
\description{
Analyze the mixed layer in GCM simulations
}
\author{
\strong{Maintainer}: Johannes Muelmenstaedt \email{johannes.muelmenstaedt@pnnl.gov}

}
