% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ne30pg2}
\alias{ne30pg2}
\title{E3SM ne30pg2 grid}
\format{
A data frame with 3 variables:
\describe{
  \item{ncol}{grid cell ("column") index}
  \item{lon}{grid cell midpoint longitude}
  \item{lat}{grid cell midpoint latitude}
  \item{area}{grid cell area}
  \item{LANDFRAC}{grid cell fractional land area}
}
}
\source{
Any E3SMv2 (other versions not tested) ne30pg2 output file
    containing a single timestep of the LANDFRAC variable (i.e.,
    nfilt=1 namelist option).  The specific file used is given
    below.

Creation:
\code{nc <- nc_open("scratch//e3sm_scratch/v2_aer_nudged/v2_1850aer_nudged/run/v2_1850aer_nudged.eam.h0.2012-12.nc");
   ne30pg2 <- dplyr::full_join(plotutils::nc.to.df(nc, c("lon", "lat", "area")),
                               plotutils::nc.to.df(nc, c("LANDFRAC")) %>% dplyr::select(-time))}

Use:
    devtools::use_data(ne30pg2, pkg = ".", internal = FALSE)}
}
\usage{
ne30pg2
}
\description{
A dataset containing the latitude, longitude, cell area, and land fraction of the E3SM ne30pg2 grid
}
\keyword{datasets}
