% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{neighbors}
\alias{neighbors}
\title{Find the nearest neighbors of each grid point}
\usage{
neighbors(df.grid, grid.spacing = 204)
}
\arguments{
\item{grid.spacing}{max distance (km) between nearest neighbors; 204 is the value for ne30pg2}
}
\description{
Find the nearest neighbors of each grid point
}
