% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{postprocess.cdnc.lwp.cloud3d}
\alias{postprocess.cdnc.lwp.cloud3d}
\title{Worth a try, but doesn't fit in memory for the 3D variables for a year}
\usage{
postprocess.cdnc.lwp.cloud3d(
  path = "~/scratch/e3sm_scratch/EAMv1_entrain_CDNC-LWP/postprocessed"
)
}
\description{
Worth a try, but doesn't fit in memory for the 3D variables for a year
}
