% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwp_utils.R
\name{predict_inverted.v}
\alias{predict_inverted.v}
\title{Use inverted v fit to predict LWP from Nd}
\usage{
predict_inverted.v(df.train, df.apply)
}
\description{
Use inverted v fit to predict LWP from Nd
}
