% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwp_utils.R
\name{predict_lwp_cdnc_conditional}
\alias{predict_lwp_cdnc_conditional}
\title{Use full P(LWP | Nd) to predict LWP from Nd}
\usage{
predict_lwp_cdnc_conditional(conditional.train, conditional.apply)
}
\arguments{
\item{conditional.train}{A data.frame produced by
lwp.cdnc.conditional() that constitutes the P(L|N) to be used
for prediction}

\item{conditional.apply}{A data.frame produced by
    lwp.cdnc.conditional that constitutes the perturbed P(N) to
    which the P(L|N) is to be applied

Note: the conditionals shall have idential cdnc_ic bins

Note: the conditionals shall not be grouped}
}
\description{
Use full P(LWP | Nd) to predict LWP from Nd
}
