% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.ccn}
\alias{process.ccn}
\title{Save PBL depth, Sc criteria, CCN}
\usage{
process.ccn(
  path = "/global/cscratch1/sd/jmuelmen/e3sm_scratch/EAMv1_entrain_NEP/run",
  pattern = "EAMv1_entrain.*cam.h1.2000-.*.nc"
)
}
\description{
Save PBL depth, Sc criteria, CCN
}
