% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.eagles.ppe}
\alias{process.eagles.ppe}
\title{Process Po-Lun's EAGLES v2 PPElet (explicit list of runs is in the source)}
\usage{
process.eagles.ppe(path = "/global/cscratch1/sd/plma/e3sm_scratch")
}
\value{
data.frame with conditional probability summaries for all
    runs; I saved these to eaglesv2_ppe.rds
}
\description{
This function can (but doesn't have to) be run as a two-stage
workflow, first producing summary files for each E3SM run, then
combining them
}
