% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.F2010A.entrain}
\alias{process.F2010A.entrain}
\title{Process 3D profiles in h9; combine with time-step output of T, Q,
CLDLIQ in h8 for tendencies; combine with ACI diagnostics in h7}
\usage{
process.F2010A.entrain(
  path = "~/scratch/e3sm_scratch/v2_F2010A0_00/run",
  pattern = ".*.eam.h9..*.nc",
  time.min = "1990-01-01",
  time.max = "2010-01-01",
  h.entrain = "eam.h9",
  h.tendenc = "eam.h8",
  h.aerocom = "eam.h7",
  parallel = TRUE
)
}
\description{
Process 3D profiles in h9; combine with time-step output of T, Q,
CLDLIQ in h8 for tendencies; combine with ACI diagnostics in h7
}
