% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.NEP.entrain}
\alias{process.NEP.entrain}
\title{Pluck the entrainment debugging variables from the grid point with
max Sc occurrence from the EAMv1_entrain_NEP run}
\usage{
process.NEP.entrain(
  path = "/qfs/people/muel306/scratch/e3sm_scratch/EAMv1_entrain_NEP_compy_nudged/run",
  pattern = "EAMv1_entrain_.*.cam.h1.20.0-0[7-9]-.*.nc",
  icol = 291
)
}
\description{
By default, only July, August, September are used
}
\details{
In some runs, the correct pattern is ...cam.h2...
}
