% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.NEP}
\alias{process.NEP}
\title{Pluck the grid point with max Sc occurrence from the EAMv1_entrain_NEP run}
\usage{
process.NEP(
  path = "/global/cscratch1/sd/jmuelmen/e3sm_scratch/EAMv1_entrain_NEP/run/",
  pattern = "EAMv1_entrain_NEP.cam.h1.2000-.*.nc",
  icol = 316
)
}
\description{
Pluck the grid point with max Sc occurrence from the EAMv1_entrain_NEP run
}
