% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.NEP.uniform.neighbors}
\alias{process.NEP.uniform.neighbors}
\title{Extract the fields needed for the entrainment calculation from all
columns satisfying the uniform PBL height among nearest neighbors
condition}
\usage{
process.NEP.uniform.neighbors(
  path = "/qfs/people/muel306/scratch/e3sm_scratch/EAMv1_entrain_NEP_compy_nudged/run",
  pattern = "EAMv1_entrain_.*.cam.h1.20.0-.*.nc",
  icol.select = 237:346
)
}
\description{
Extract the fields needed for the entrainment calculation from all
columns satisfying the uniform PBL height among nearest neighbors
condition
}
