% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.pblh.inversion}
\alias{process.pblh.inversion}
\title{Take a look at different PBL height definitions}
\usage{
process.pblh.inversion(
  path = "/qfs/people/muel306/scratch/e3sm_scratch/EAMv1_entrain_NEP_compy_nudged/run",
  pattern = "EAMv1_entrain_.*.cam.h1.20.0-.*.nc"
)
}
\description{
Take a look at different PBL height definitions
}
