% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm_decomposition.R
\name{process.regression}
\alias{process.regression}
\title{process a list of file names, extracting the information necessary
for dalpha/dLWP}
\usage{
process.regression(lf, icc.max = 0.02, time.min, time.max)
}
\description{
process a list of file names, extracting the information necessary
for dalpha/dLWP
}
