% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process.Sc}
\alias{process.Sc}
\title{Pluck the profiles that meet Sc critera from the EAMv1_entrain_CDNC-LWP run}
\usage{
process.Sc(
  path = "/global/cscratch1/sd/jmuelmen/e3sm_scratch/EAMv1_entrain_CDNC-LWP/run",
  pattern = "EAMv1_entrain_CDNC-LWP.cam.h1.2010-.*.nc"
)
}
\description{
Pluck the profiles that meet Sc critera from the EAMv1_entrain_CDNC-LWP run
}
