% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{process_e3sm_aer}
\alias{process_e3sm_aer}
\title{Process a single run into Nd-LWP histograms}
\usage{
process_e3sm_aer(
  path = "/compyfs/dtn/muel306/EAMv1_nudged_2010/run",
  pattern = "EAMv1_nudged_2010.cam.h2..*.nc",
  time.min = "2010-01-01",
  time.max = "2015-01-01",
  version = 1
)
}
\value{
none; data.frame with conditional probability summaries is
    saved to file
}
\description{
Process a single run into Nd-LWP histograms
}
