% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{reconstruct_inversion}
\alias{reconstruct_inversion}
\title{Match gradients above and below the inversion to provide PBL height}
\usage{
reconstruct_inversion(p, z3, x, debug.plot = FALSE)
}
\description{
Match gradients above and below the inversion to provide PBL height
}
