% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{reconstruct_inversion_pressure}
\alias{reconstruct_inversion_pressure}
\title{Match gradients above and below the inversion to provide PBL height}
\usage{
reconstruct_inversion_pressure(p, z3, x, debug.plot = FALSE)
}
\description{
Same as reconstruct_inversion(), but returns PBL top pressure.
}
\details{
For mass budgets, this function is more exact than
reconstruct_inversion(), which fails to account for density
variation within the inversion "zone of confusion".
}
