% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_e3sm.R
\name{regionalize.F2010A.entrain}
\alias{regionalize.F2010A.entrain}
\title{Select Sc regions from the output of process.F2010A.entrain()}
\usage{
regionalize.F2010A.entrain(file = "~/scratch/", grid = "~/scratch/ne30pg2.rds")
}
\description{
Select Sc regions from the output of process.F2010A.entrain()
}
