% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regionalize_annual}
\alias{regionalize_annual}
\title{Define subsidence Sc region by location only (annual-mean
occurrence frequency of Medeiros and Stevens 2011 criteria)}
\usage{
regionalize_annual(lon, lat, fscu.annual, landfrac)
}
\description{
Define subsidence Sc region by location only (annual-mean
occurrence frequency of Medeiros and Stevens 2011 criteria)
}
