% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regionalize_seasonal}
\alias{regionalize_seasonal}
\title{Define subsidence Sc region by season and location (annual-mean
occurrence frequency of Medeiros and Stevens 2011 criteria)}
\usage{
regionalize_seasonal(month, lon, lat, fscu.annual, landfrac)
}
\description{
Define subsidence Sc region by season and location (annual-mean
occurrence frequency of Medeiros and Stevens 2011 criteria)
}
