% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regionalize_zhang}
\alias{regionalize_zhang}
\title{Define subsidence Sc region by location only (Zhang and Feingold
2023 lon-lat boxes)}
\usage{
regionalize_zhang(lon, lat)
}
\description{
Define subsidence Sc region by location only (Zhang and Feingold
2023 lon-lat boxes)
}
