% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3sm_utils.R
\name{replace_with_conventional_units}
\alias{replace_with_conventional_units}
\title{Replace with conventional LWP and Nd units}
\usage{
replace_with_conventional_units(df)
}
\description{
Replace with conventional LWP and Nd units
}
